Extract generate date function
