+++ src/mock/helpers.js
@@ -1,3 +1,5 @@
+import dayjs from "dayjs";
+
 // Функция из интернета по генерации случайного числа из диапазона
 // Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
 export const getRandomInteger = (a = 0, b = 1) => {
@@ -12,3 +14,19 @@ export const getRandomInteger = (a = 0, b = 1) => {
 // Для "продуктового" кода используйте что-то понадежнее,
 // вроде nanoid - https://github.com/ai/nanoid
 export const generateId = () => Date.now() + parseInt(Math.random() * 10000, 10);
+
+export const generateDate = () => {
+  const year = getRandomInteger(0, 3);
+  const month = getRandomInteger(0, 3);
+  const day = getRandomInteger(0, 6);
+  const hour = getRandomInteger(0, 12);
+  const min = getRandomInteger(0, 30);
+
+  return dayjs()
+    .subtract(year, `year`)
+    .subtract(month, `month`)
+    .subtract(day, `day`)
+    .subtract(hour, `hour`)
+    .subtract(min, `minute`)
+    .toDate();
+};
